[hir2mpl]Internal commit msg:     [hir2mpl] cleancode2     [hir2mpl]fix mulalign offset mistake     [hir2mpl] fix struct field offset in boundary check     [hir2mpl]fix mulattraligned mistake     [hir2mpl] add initlist opt     [ICSL] Clean Code 53     [ICSL] Clean Code 52     [hir2mpl] fix lto inline func name mangled     [ci] clean script code 03     [hir2mpl] fix 128 bug if use npe check option     update to llvm 15     [DTS2023051811106 fix maple unorderedmap memLeak MapleUnorderedMap<MapleString*>]     [driver]--help     [ICSL] Clean Code 50     [me] struct field offset     [codecheck]     [hir2mpl] fix if use lto and safe-region at the same time [mapleall]Internal commit msg:     [mapleall] cleancode2  53     [mplcg]change tls desc     cg_bugfix_fix_last_insn_in_ra     cg_bugfix_fix_EliminateSpecifcSXTPattern     Fix unknown ill code bug     [mapleall] cleancode2  52     [codecheck delete const for move CSTR]     [codecheck me&utils]     add nullptr check before dereference     [mplcg]fix bug in PGO (filp op)     [codecheck me_merge_stmts]     [codecheck util]     [codecheck util]     [ICSL] Clean Code 55     [driver]-o-     [DTS2023052908766 asan tls Anchor nullptr]     [mapleall] cleancode2  51     [DTS2023052706768]     [mapleall] cleancode2  50     [maplecg] add vra register use infomation for propInmem during cgtargetprop phase     [mapleall] cleancode2  49     [mplcg] re-code NeedGOT logic     [mplcg] Csel with RZR to condset in peephole     force add realocc without check same symbol     [mapleall] cleancode2  48     cg_bugfix_fix_dbg_insn_id_priority     cg_bugfix_remove_tmp_dir_and_runtime_info_in_release_build     [ICSL] Clean Code 54     [codecheck me]     [mapleall] bugfix compile err in debug mode     [hir2mpl] add initlist opt     [ICSL] Clean Code 52     [driver]xlinker     use copy-assign instead of move-assign when copy Intval     redundant uxth     [mapleall] cleancode2  47     [mapleall] reduce compile time     [hir2mpl] fix lto inline func name mangled     [mapleall] bugfix vrp     [ci] clean script code 03     enhance condition cvt to get better opt     Revert "LTO"     [litepgo] Bugfix for hpf pgouse with whitelist     [mplcg] Enhance CmpBranchesPattern in validbitopt     update to llvm 15     [mapleall] cleancode2  46     [codecheck cg & driver]     [mplcg] CondSet optimization in ico     [ci] clean script code 01     [DTS2023051811106 fix maple unorderedmap memLeak MapleUnorderedMap<MapleString*>]     Fix long double ZeroOpnd assignment     skip implicit cvt     [mapleall] bugfix sra     [driver]--help     [ICSL] Clean Code 48     [mpl2mpl]Bugfix for invalid memorder     [codecheck]     [ICSL] Clean Code 51     [mapleall] cleancode2  44     [mapleall] cleancode2  43     [DTS2023052215052]     [DTS2023051811262 fix mapleVector memLeak]     [codecheck cg simp]     [abi] hfa/hva with struct align bugfix     [codecheck cg driver]     [ICSL] Clean Code 49     Fix do-while processing in expand128float     [ICSL] Clean Code 47     [mapleall] Inline bugfix     [mplcg] update pgo profile     [atomic]Handle invalid memorder param by referring to gcc     [ICSL] Clean Code 46     [ICSL] revert cc42 change of bb.h for gmdb bug     [me] struct field offset     [mapleall] cleancode2  45     cg_bugfix_always_emit_label_for_first_bb     [litepgo] Bugfix for cfigen after pgouse     cg_bugfix_fix_float_const_label_name     [mplme] enhance fold branch of cfgo     [mapleall] cleancode2  42     [ICSL] Clean Code 42     frame refactor     [ICSL] Clean Code 45     [mapleall] bugfix objsize case timeout     [codecheck cg]     [mapleall] cleancode2  41     [simplify] get sub field recurrsively     keep intmin calculation order     [codecheck]     [ICSL] Clean Code 44     [litepgo] Bugfix "pred of critical edge bb do not goto cbb"     [driver]werror     [mplcg] fix bug in inter-dependency-analysis     [ICSL] Clean Code 43     check pred num     [cg] for phi insn, live only set live out from source BB and ico enhanced     [maplecg] Add AndAndCmpBranchesToTstPattern in PeepHole     [driver] fix maple -h and add lto ut     cg_bugfix_constant_pool_for_float     [driver]fake .a as input     LTO     [maple_san]disable maplesan build     [driver]-O2 -D_FORTIFY_SOURCE -O0     cg_bugfix_add_imm_adjust_for_ubfx     !1305 AddressSanitizer for OpenArkCompiler     [cg] ra spill with remat opnd bugfix     [mapleall] bugfix globalopt     [ICSL] Clean Code 41     [mapleall] bugfix scev     [mapleall] add new phase to opt continuous memory concatenation     [mplcg] clean code for scheduling     [mplcg] Bugfix for DTS2023051206585     [mapleall] enhance loopunrolling with cvt opnd     [mapleall] support OP_and in vrp     [mapleall] simplify OP_sub in irmap     [ICSL] Clean Code 40     [ICSL] Clean Code 39     [mplcg] Add UbfxAndCbzToTbzPattern to NormalPattern     [driver]fixbug operate nullptr     [maplcg] fix bug in scheduling inline-asm     [slp] reduce mempool alloc during slp     cg_bugfix_tls_fix     [codecheck] loop codecheck fix     [codecheck aarch64_aggressive_opt]     [mplcg] tuning scheduling     do not trans (a + min) to (a - min), the latter is not safe     [driver]enable -specs     [DTS2023042515189]     cg_bugfix_fix_debug_info_in_opt     [cg] loop analysis reduce complie time     [cg]Bugfix for DTS2023042704792     [mapleall] Enhance pgouse bblayout for cold path     [cg] add pattern to remove uxth     cg_bugfix_fix_ico_morepredspattern     [mapleall] bugfix ssaupdate in vrp     [mapleall] bugfix vrp     [mapleall] bugfix delete bb in vrp     [mapleall] bugfix bblayout when create new bb     [mapleall] bugfix integer overflow in vrp     [ICSL] Clean Code 38     [litepgo] Bugfix for DTS2023050601491     [mplcg]pgo DTS bugfix

+++ src/mapleall/maple_be/include/cg/aarch64/aarch64_mop_split.h
@@ -15,6 +15,12 @@
 #ifndef MAPLEBE_INCLUDE_CG_AARCH64_AARCH64_MOP_SPLIT_H
 #define MAPLEBE_INCLUDE_CG_AARCH64_AARCH64_MOP_SPLIT_H
 
+#include "aarch64_cg.h"
+#include "aarch64_isa.h"
+#include "cg_irbuilder.h"
+#include "cgbb.h"
+#include "common_utils.h"
+
 namespace maplebe {
 // Supply a new reg operand for insn split process, which type is kRegTyInt for immediate.
 // Before regalloc: create a new virtual reg;
@@ -312,7 +318,8 @@ inline void AddSubWithLslSplit(Insn *insn, bool isAdd, bool is64Bits, bool isAft
     return;
   }
   BB *bb = insn->GetBB();
-  ImmOperand &newImmOpnd = opndBuilder->CreateImm(size, static_cast<uint64>(immOpnd.GetValue()) << k12BitSize);
+  ImmOperand &newImmOpnd = opndBuilder->CreateImm(size, static_cast<int64>(static_cast<uint64>(immOpnd.GetValue()) <<
+                                                                           k12BitSize));
   MOperator mOpCode = is64Bits ? (isAdd ? MOP_xaddrri12 : MOP_xsubrri12) : (isAdd ? MOP_waddrri12 : MOP_wsubrri12);
   Insn &newInsn = insnBuilder->BuildInsn(mOpCode, insn->GetOperand(kInsnFirstOpnd),
       insn->GetOperand(kInsnSecondOpnd), newImmOpnd);
@@ -391,7 +398,9 @@ inline void CondCompareInsnSplit(Insn *insn, bool is64Bits, bool isAfterRegAlloc
 inline void MOP_wmovri32Split(Insn *curInsn, bool /* isAfterRegAlloc */, InsnBuilder *insnBuilder,
     OperandBuilder *opndBuilder) {
   // If higher 32bits of immVal have 1, we will truncate and keep lower 32 bits.
-  int64 immVal = (static_cast<ImmOperand&>(curInsn->GetOperand(kInsnSecondOpnd)).GetValue()) & 0x00000000FFFFFFFFULL;
+  int64 immVal = static_cast<int64>(
+      static_cast<uint64>(static_cast<ImmOperand&>(curInsn->GetOperand(kInsnSecondOpnd)).GetValue()) &
+      0x00000000FFFFFFFFULL);
   ImmOperand &immOpnd = opndBuilder->CreateImm(k64BitSize, immVal, true);
   curInsn->SetOperand(kInsnSecondOpnd, immOpnd);
   if (curInsn->VerifySelf()) { return; }
